requirements.txt
websocket-client==1.7.0
kafka-python==2.0.2
pydantic==2.5.3
python-dotenv==1.0.0
structlog==24.1.0
prometheus-client==0.19.0
requests==2.31.0
